Time: 17 ms (65.16%), Space: 43.7 MB (81.55%) - LeetHub

+++ 0018-4sum/0018-4sum.java
@@ -1,8 +1,7 @@
 public class Solution {
 	
     public static List<List<Integer>> fourSum(int[] nums, int target) {
-		List<List<Integer>> result = new ArrayList<>();
-
+        List<List<Integer>> result = new ArrayList<>();
 		Arrays.sort(nums);
 
 		for (int i = 0; i < nums.length - 3; i++) {
@@ -25,7 +24,7 @@ public class Solution {
 				int l = nums.length - 1;
 				while (k < l) {
 
-					int sum = nums[i] + nums[j] + nums[k] + nums[l];
+					long sum = nums[i] + nums[j] + nums[k] + nums[l];
 
 					if (sum == target) {
 						result.add(Arrays.asList(nums[i], nums[j], nums[k], nums[l]));
